refactor: Make update device adminstate cmd shorter
Change update device adminState command from:
./edgex-cli device adminstate update -d -s
to
./edgex-cli device adminstate -d -s
Closes: https://github.com/edgexfoundry/edgex-cli/issues/309
